backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
psycopg2-binary==2.9.9
